Refactor screen logic structure
