+++ zenhabits_app/lib/main.dart
@@ -7,11 +7,11 @@ import 'package:zenhabits_app/domain/usecases/insert_habit_usecase.dart';
 import 'package:zenhabits_app/domain/usecases/delete_habit_usecase.dart';
 import 'package:zenhabits_app/domain/usecases/get_habits_usecase.dart';
 import 'package:zenhabits_app/domain/usecases/update_habit_usecase.dart';
-import 'package:zenhabits_app/presentation/screens/create_habit_screen.dart';
-import 'package:zenhabits_app/presentation/screens/home_screen.dart';
-import 'package:zenhabits_app/presentation/screens/login_screen.dart';
+import 'package:zenhabits_app/presentation/screens/create/habit/create_habit_screen.dart';
+import 'package:zenhabits_app/presentation/screens/home/home_screen.dart';
+import 'package:zenhabits_app/presentation/screens/login/login_screen.dart';
 import 'package:zenhabits_app/presentation/viewmodels/habit_view_model.dart';
-import 'package:zenhabits_app/presentation/screens/signup_screen.dart';
+import 'package:zenhabits_app/presentation/screens/signup/signup_screen.dart';
 import 'package:zenhabits_app/data/repositories/users_repository.dart';
 import 'package:zenhabits_app/domain/usecases/insert_user_usecase.dart';
 import 'package:zenhabits_app/domain/usecases/get_user_usecase.dart';
