+++ zenhabits_app/lib/presentation/screens/login/login_state.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class LoginState {
+  final TextEditingController usernameController = TextEditingController();
+  final TextEditingController passwordController = TextEditingController();
+
+  void dispose() {
+    usernameController.dispose();
+    passwordController.dispose();
+  }
+}
