+++ zenhabits_app/lib/presentation/screens/signup/signup_form.dart
@@ -1,0 +1,171 @@
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:zenhabits_app/presentation/viewmodels/user_view_model.dart';
+import 'signup_controller.dart';
+
+class SignUpForm extends StatefulWidget {
+  const SignUpForm({super.key});
+
+  @override
+  State<SignUpForm> createState() => _SignUpFormState();
+}
+
+class _SignUpFormState extends State<SignUpForm> {
+  final _formKey = GlobalKey<FormState>();
+  final _usernameController = TextEditingController();
+  final _emailController = TextEditingController();
+  final _passwordController = TextEditingController();
+  final _confirmPasswordController = TextEditingController();
+
+  @override
+  void dispose() {
+    _usernameController.dispose();
+    _emailController.dispose();
+    _passwordController.dispose();
+    _confirmPasswordController.dispose();
+    super.dispose();
+  }
+
+  void _submitForm() async {
+    if (!_formKey.currentState!.validate()) return;
+
+    final userViewModel = context.read<UserViewModel>();
+    final controller = SignUpController(
+      userViewModel: userViewModel,
+      context: context,
+    );
+
+    await controller.signUp(
+      username: _usernameController.text.trim(),
+      email: _emailController.text.trim(),
+      password: _passwordController.text,
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final isLoading = context.watch<UserViewModel>().isLoading.value;
+
+    return Form(
+      key: _formKey,
+      child: Column(
+        children: [
+          const SizedBox(height: 20),
+          const Text(
+            'CREAR CUENTA NUEVA',
+            style: TextStyle(
+              fontWeight: FontWeight.bold,
+              fontSize: 24,
+              color: Colors.brown,
+              decoration: TextDecoration.underline,
+            ),
+          ),
+          const SizedBox(height: 30),
+          _buildTextField(
+            controller: _usernameController,
+            hint: 'username',
+            validator: (value) =>
+                value!.isEmpty ? 'Ingrese un nombre de usuario' : null,
+          ),
+          const SizedBox(height: 16),
+          _buildTextField(
+            controller: _emailController,
+            hint: 'email',
+            keyboardType: TextInputType.emailAddress,
+            validator: (value) {
+              if (value == null || value.isEmpty) return 'Ingrese su correo electrónico';
+              if (!value.contains('@') || !value.contains('.')) return 'Correo inválido';
+              return null;
+            },
+          ),
+          const SizedBox(height: 16),
+          _buildTextField(
+            controller: _passwordController,
+            hint: 'contraseña',
+            obscureText: true,
+            validator: (value) =>
+                value!.length < 6 ? 'Mínimo 6 caracteres' : null,
+          ),
+          const SizedBox(height: 16),
+          _buildTextField(
+            controller: _confirmPasswordController,
+            hint: 'confirmar contraseña',
+            obscureText: true,
+            validator: (value) =>
+                value != _passwordController.text ? 'No coinciden' : null,
+          ),
+          const SizedBox(height: 30),
+          Row(
+            children: [
+              Expanded(
+                child: ElevatedButton(
+                  onPressed: isLoading ? null : _submitForm,
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: Colors.orange,
+                    padding: const EdgeInsets.symmetric(vertical: 14),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(12),
+                    ),
+                  ),
+                  child: isLoading
+                      ? const CircularProgressIndicator(color: Colors.white)
+                      : const Text(
+                          'ACEPTAR',
+                          style: TextStyle(
+                            fontWeight: FontWeight.bold,
+                            color: Colors.white,
+                          ),
+                        ),
+                ),
+              ),
+              const SizedBox(width: 16),
+              Expanded(
+                child: ElevatedButton(
+                  onPressed: () => Navigator.pop(context),
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: const Color(0xFFFFC66B),
+                    padding: const EdgeInsets.symmetric(vertical: 14),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(12),
+                    ),
+                  ),
+                  child: const Text(
+                    'CANCELAR',
+                    style: TextStyle(
+                      fontWeight: FontWeight.bold,
+                      color: Colors.white,
+                    ),
+                  ),
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildTextField({
+    required TextEditingController controller,
+    required String hint,
+    bool obscureText = false,
+    TextInputType? keyboardType,
+    String? Function(String?)? validator,
+  }) {
+    return TextFormField(
+      controller: controller,
+      obscureText: obscureText,
+      keyboardType: keyboardType,
+      validator: validator,
+      decoration: InputDecoration(
+        hintText: hint,
+        filled: true,
+        fillColor: Colors.white,
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide.none,
+        ),
+      ),
+    );
+  }
+}
